Rename duplicate .cpp files.
