+++ src/ripple/overlay/make_Overlay.h
@@ -22,7 +22,7 @@
 
 #include <ripple/server/ServerHandler.h>
 #include <ripple/overlay/Overlay.h>
-#include <ripple/resource/Manager.h>
+#include <ripple/resource/ResourceManager.h>
 #include <ripple/basics/Resolver.h>
 #include <beast/threads/Stoppable.h>
 #include <beast/module/core/files/File.h>
