+++ src/ripple/peerfinder/impl/Bootcache.h
@@ -20,7 +20,7 @@
 #ifndef RIPPLE_PEERFINDER_BOOTCACHE_H_INCLUDED
 #define RIPPLE_PEERFINDER_BOOTCACHE_H_INCLUDED
 
-#include <ripple/peerfinder/Manager.h>
+#include <ripple/peerfinder/PeerfinderManager.h>
 #include <ripple/peerfinder/impl/Store.h>
 #include <beast/utility/Journal.h>
 #include <beast/utility/PropertyStream.h>
