+++ src/ripple/peerfinder/impl/Endpoint.cpp
@@ -18,7 +18,7 @@
 //==============================================================================
 
 #include <BeastConfig.h>
-#include <ripple/peerfinder/Manager.h>
+#include <ripple/peerfinder/PeerfinderManager.h>
 
 namespace ripple {
 namespace PeerFinder {
