+++ src/ripple/peerfinder/impl/SlotImp.h
@@ -21,7 +21,7 @@
 #define RIPPLE_PEERFINDER_SLOTIMP_H_INCLUDED
 
 #include <ripple/peerfinder/Slot.h>
-#include <ripple/peerfinder/Manager.h>
+#include <ripple/peerfinder/PeerfinderManager.h>
 #include <beast/config/CompilerConfig.h>
 #include <beast/container/aged_unordered_map.h>
 #include <beast/container/aged_container_utility.h>
