+++ src/ripple/peerfinder/impl/Source.h
@@ -20,7 +20,7 @@
 #ifndef RIPPLE_PEERFINDER_SOURCE_H_INCLUDED
 #define RIPPLE_PEERFINDER_SOURCE_H_INCLUDED
 
-#include <ripple/peerfinder/Manager.h>
+#include <ripple/peerfinder/PeerfinderManager.h>
 #include <beast/smart_ptr/SharedObject.h>
 #include <boost/system/error_code.hpp>
