+++ src/ripple/resource/ResourceManager.h
