+++ src/ripple/test/jtx.h
@@ -29,7 +29,7 @@
 #include <ripple/test/jtx/Env.h>
 #include <ripple/test/jtx/fee.h>
 #include <ripple/test/jtx/flags.h>
-#include <ripple/test/jtx/json.h>
+#include <ripple/test/jtx/jtx_json.h>
 #include <ripple/test/jtx/JTx.h>
 #include <ripple/test/jtx/memo.h>
 #include <ripple/test/jtx/multisign.h>
@@ -54,4 +54,3 @@
 #include <ripple/test/jtx/utility.h>
 
 #endif
-
