+++ src/ripple/unity/test.cpp
@@ -26,7 +26,7 @@
 #include <ripple/test/jtx/impl/Env.cpp>
 #include <ripple/test/jtx/impl/fee.cpp>
 #include <ripple/test/jtx/impl/flags.cpp>
-#include <ripple/test/jtx/impl/json.cpp>
+#include <ripple/test/jtx/impl/jtx_json.cpp>
 #include <ripple/test/jtx/impl/memo.cpp>
 #include <ripple/test/jtx/impl/multisign.cpp>
 #include <ripple/test/jtx/impl/offer.cpp>
