+++ src/ripple/unity/validators.cpp
@@ -20,6 +20,6 @@
 #include <BeastConfig.h>
 #include <ripple/validators/impl/ConnectionImp.cpp>
 #include <ripple/validators/impl/Logic.cpp>
-#include <ripple/validators/impl/Manager.cpp>
+#include <ripple/validators/impl/ValidatorManager.cpp>
 #include <ripple/validators/impl/StoreSqdb.cpp>
 #include <ripple/validators/tests/Validators.test.cpp>
